service/requirements.txt
requests==2.24.0
hvac==0.10.4
kubernetes==11.0.0
pyjwt==1.7.1